requirements.txt
Flask>=2.2
requests>=2.28
beautifulsoup4>=4.12
python-dotenv>=1.0
tqdm>=4.65
numpy>=1.24
scikit-learn>=1.3
sentence-transformers>=2.2
openai>=1.0.0
